# openapi.yaml
openapi: 3.0.3
info:
  title: RouteSort API
  description: |
    Production-ready API for the RouteSort waste pickup optimization system.
    Provides endpoints for pickup requests, driver management, route optimization,
    and image classification for recycling guidance.
  version: 1.0.0
  contact:
    name: RouteSort Team
  license:
    name: ISC
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://routesort.vercel.app/api
    description: Production server
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token
  schemas:
    Coordinate:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
        lng:
          type: number
          format: double
          minimum: -180
          maximum: 180
...
    Driver:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        vehicle_number:
          type: string
        vehicle_type:
          type: string
        capacity_kg:
          type: integer
        status:
          type: string
          enum: [available, on_route, offline]
        current_latitude:
          type: number
          format: double
        current_longitude:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
...
    OptimizeRouteRequest:
      type: object
      required: [driver_ids, pickup_ids]
      properties:
        driver_ids:
          type: array
          items:
            type: integer
        pickup_ids:
          type: array
          items:
            type: integer
        vehicle_capacities:
          type: object
          additionalProperties:
            type: integer
        max_route_length:
          type: number
          description: Maximum route length in km
        priority_weight:
          type: number
          description: Weight for priority pickups (0-1)
    RouteStop:
      type: object
      properties:
        pickup_id:
          type: integer
        seq:
          type: integer
        eta:
          type: string
          format: date-time
    OptimizedRoute:
      type: object
      properties:
        driver_id:
          type: integer
        path:
          type: array
          items:
            $ref: '#/components/schemas/Coordinate'
        stops:
          type: array
          items:
            $ref: '#/components/schemas/RouteStop'
        total_distance:
          type: number
          description: Total distance in km
        estimated_duration:
          type: number
          description: Estimated duration in minutes
    OptimizeRouteResponse:
      type: object
      properties:
        routes:
          type: array
          items:
            $ref: '#/components/schemas/OptimizedRoute'
        stats:
          type: object
          properties:
            total_pickups:
              type: integer
            total_distance:
              type: number
            average_route_length:
              type: number
            unassigned_pickups:
              type: array
              items:
                type: integer
...
paths:
...
  /drivers:
    get:
      tags: [drivers]
      summary: List all drivers
      description: Returns list of drivers with their current status and location
      security:
        - ClerkAuth: []
      responses:
        '200':
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Driver'
              example:
                - id: 1
                  user_id: 10
                  vehicle_number: "TRK-001"
                  vehicle_type: "compactor"
                  capacity_kg: 5000
                  status: "available"
                  current_latitude: 40.7580
                  current_longitude: -73.9855
...
  /routes/optimize:
    post:
      tags: [routes]
      summary: Optimize routes for drivers
      description: |
        Runs CVRP algorithm to optimize pickup routes for specified drivers.
        Uses greedy + 2-opt heuristic with capacity constraints.
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizeRouteRequest'
            example:
              driver_ids: [1, 2, 3]
              pickup_ids: [10, 11, 12, 13, 14, 15]
              vehicle_capacities:
                "1": 5000
                "2": 4000
                "3": 6000
              max_route_length: 50
              priority_weight: 0.3
      responses:
        '200':
          description: Optimized routes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizeRouteResponse'
              example:
                routes:
                  - driver_id: 1
                    path:
                      - lat: 40.7128
                        lng: -74.0060
                      - lat: 40.7580
                        lng: -73.9855
                    stops:
                      - pickup_id: 10
                        seq: 1
                        eta: "2024-01-20T09:00:00Z"
                      - pickup_id: 13
                        seq: 2
                        eta: "2024-01-20T09:30:00Z"
                    total_distance: 15.5
                    estimated_duration: 45
                stats:
                  total_pickups: 6
                  total_distance: 45.2
                  average_route_length: 15.1
                  unassigned_pickups: []
...
security:
  - ClerkAuth: []